requirements.txt
langchain
elasticsearch
python-dotenv
datasets
openai
tiktoken
sentence-transformers